[OPENIDM-2058][CR-4593] Fixed response code/content issues with Configuration Service REST API.

+++ openidm-config/src/main/java/org/forgerock/openidm/config/manage/ConfigObjectService.java
@@ -42,6 +42,7 @@ import org.apache.felix.scr.annotations.Reference;
 import org.apache.felix.scr.annotations.ReferencePolicy;
 import org.apache.felix.scr.annotations.Service;
 import org.forgerock.json.fluent.JsonValue;
+import org.forgerock.json.fluent.JsonValueException;
 import org.forgerock.json.resource.ActionRequest;
 import org.forgerock.json.resource.BadRequestException;
 import org.forgerock.json.resource.ConflictException;
@@ -135,15 +136,15 @@ public class ConfigObjectService implements RequestHandler {
      * @throws ForbiddenException  if access to the object is forbidden.
      * @throws BadRequestException if the passed identifier is invalid
      */
-    public Map<String, Object> read(ResourceName resourceName) throws ResourceException {
+    public JsonValue read(ResourceName resourceName) throws ResourceException {
         logger.debug("Invoking read {}", resourceName.toString());
-        Map<String, Object> result = null;
+        JsonValue result = null;
 
         try {
 
             if (resourceName.isEmpty()) {
                 // List all configurations
-                result = new HashMap<String, Object>();
+                result = new JsonValue(new HashMap<String, Object>());
                 Configuration[] rawConfigs = configAdmin.listConfigurations(null);
                 List configList = new ArrayList();
                 if (null != rawConfigs) {
@@ -177,8 +178,7 @@ public class ConfigObjectService implements RequestHandler {
                 }
                 Dictionary props = config.getProperties();
                 JSONEnhancedConfig enhancedConfig = new JSONEnhancedConfig();
-                JsonValue value = enhancedConfig.getConfiguration(props, resourceName.toString(), false);
-                result = value.asMap();
+                result =  enhancedConfig.getConfiguration(props, resourceName.toString(), false);
                 logger.debug("Read configuration for service {}", resourceName);
             }
         } catch (ResourceException ex) {
@@ -251,6 +251,9 @@ public class ConfigObjectService implements RequestHandler {
             logger.debug("Created new configuration for {} with {}", parsedId.toString(), dict);
         } catch (ResourceException ex) {
             throw ex;
+        } catch (JsonValueException ex) {
+            logger.warn("Invalid configuration provided for {}:" + ex.getMessage(), resourceName.toString(), ex);
+            throw new BadRequestException("Invalid configuration provided for " + resourceName.toString() + ": " + ex.getMessage(), ex);
         } catch (WaitForMetaData ex) {
             logger.info("No meta-data provider available yet to create and encrypt configuration for {}, retry later.", parsedId.toString(), ex);
             throw new InternalServerErrorException("No meta-data provider available yet to create and encrypt configuration for "
@@ -303,9 +306,9 @@ public class ConfigObjectService implements RequestHandler {
         else if (resourceName.size() > 2) {
             throw new BadRequestException("The passed identifier to update has more than two parts");
         }
-
-        try {
-            ParsedId parsedId = new ParsedId(resourceName);
+        
+        ParsedId parsedId = new ParsedId(resourceName);
+        try {    
             Configuration config = findExistingConfiguration(parsedId);
 
             Dictionary existingConfig = (config == null ? null : config.getProperties());
@@ -317,6 +320,13 @@ public class ConfigObjectService implements RequestHandler {
             logger.debug("Updated existing configuration for {} with {}", resourceName.toString(), existingConfig);
         } catch (ResourceException ex) {
             throw ex;
+        } catch (JsonValueException ex) {
+            logger.warn("Invalid configuration provided for {}:" + ex.getMessage(), resourceName.toString(), ex);
+            throw new BadRequestException("Invalid configuration provided for " + resourceName.toString() + ": " + ex.getMessage(), ex);
+        } catch (WaitForMetaData ex) {
+            logger.info("No meta-data provider available yet to update and encrypt configuration for {}, retry later.", parsedId.toString(), ex);
+            throw new InternalServerErrorException("No meta-data provider available yet to create and encrypt configuration for "
+                    + parsedId.toString() + ", retry later.", ex);
         } catch (Exception ex) {
             logger.warn("Failure to update configuration for {}", resourceName.toString(), ex);
             throw new InternalServerErrorException("Failure to update configuration for " + resourceName.toString() + ": " + ex.getMessage(), ex);
@@ -328,8 +338,7 @@ public class ConfigObjectService implements RequestHandler {
             ResultHandler<Resource> handler) {
         try {
             String rev = request.getRevision();
-            delete(request.getResourceNameObject(), rev);
-            Resource resource = new Resource(request.getResourceName(), null, new JsonValue(null));
+            Resource resource = new Resource(request.getResourceName(), null, delete(request.getResourceNameObject(), rev));
             handler.handleResult(resource);
         } catch (ResourceException e) {
             handler.handleError(e);
@@ -343,25 +352,33 @@ public class ConfigObjectService implements RequestHandler {
      *
      * @param resourceName the identifier of the resource to be deleted.
      * @param rev    the version of the object to delete or {@code null} if not provided.
+     * @return the deleted object.
      * @throws NotFoundException           if the specified object could not be found.
      * @throws ForbiddenException          if access to the object is forbidden.
      * @throws ConflictException           if version is required but is {@code null}.
      * @throws PreconditionFailedException if version did not match the existing object in the set.
      */
-    public void delete(ResourceName resourceName, String rev) throws ResourceException {
+    public JsonValue delete(ResourceName resourceName, String rev) throws ResourceException {
         logger.debug("Invoking delete configuration {} {}", resourceName.toString(), rev);
         if (resourceName.isEmpty()) {
             throw new BadRequestException("The passed identifier to delete is null");
         }
         try {
             Configuration config = findExistingConfiguration(new ParsedId(resourceName));
-
+            if (config == null) {
+                throw new NotFoundException("No existing configuration found for " + resourceName.toString() + ", can not delete the configuration.");
+            }
+            
             Dictionary existingConfig = config.getProperties();
+            JSONEnhancedConfig enhancedConfig = new JSONEnhancedConfig();
+            JsonValue value = enhancedConfig.getConfiguration(existingConfig, resourceName.toString(), false);
+
             if (existingConfig == null) {
                 throw new NotFoundException("No existing configuration found for " + resourceName.toString() + ", can not delete the configuration.");
             }
             config.delete();
             logger.debug("Deleted configuration for {}", resourceName.toString());
+            return value;
         } catch (ResourceException ex) {
             throw ex;
         } catch (Exception ex) {
